urldecode wifi access point ssid and password (#61)
* More explicit setting of wifi power mode

Use cyw43_pm_value to set the wifi power mode. The call results in the
exact same value as before (0xa11140), but makes it more explicit what
is happening. It also makes it easier to adjust the power management in
the future.

* urldecode wifi ssid and password

Allows special characters to be present in the ssid and password

---------

Co-authored-by: Shiny Mira <mira@latexland.invalid>